feat(pix): put surplus shares on the wire behind a single param encoding
PIX negotiates three per-session dimensions but only two ever travelled. The
surplus is what decides how much data an Exit delivers beyond the quota it
charged for — `transport/hopr/src/config.rs` said so outright: "it is what the
peer never learns: only the part size goes on the wire, so the surplus is
unpriced".

The two that did travel were bit-packed by hand in two mutually inconsistent
shapes — `polys << 16 | threshold` in the `SsaRequest` params word, and
`polys << 48 | threshold << 32` in `StartInitiation.additional_data` — plus a
third time across 22 test sites. Neither wire decode validated anything; the
ranges were five separate inline literals.

Narrowing the threshold to `2..=255` frees the byte the surplus needs, and both
packings collapse into `PixParams` in `hopr-protocol-pix`, the lowest common
ancestor of the two sites. It owns `to_u32`/`try_from_u32` and
`into_additional_data`/`try_from_additional_data`, so the layout exists once and
unpacking is what enforces the ranges. Threshold and surplus become `u8`
end-to-end, which makes the upper bound unrepresentable rather than checked.

`SsaDimensions` deliberately stays two fields: it is what an API caller asks
for, and the surplus is a property of the node's installed generator, not of
the request. `new_session` fills it in from there.

`SsaServerCommitmentMessage.params` stays a raw `u32` with a `dimensions()`
accessor rather than becoming a typed field. Validating inside
`StartProtocol::decode` would regress 45fc808de4: `dispatch_message` drops
undecodable packets without replying, so out-of-range params would become a
kill-switch timeout instead of the `UnacceptablePixParams` refusal they get
today. The Entry now compares the whole triple rather than the quota product,
because the surplus is not in the quota at all.

`START_PROTOCOL_VERSION` stays at 0x03: master is still 0x02 and has no
`SsaRequest` message, so no deployed peer speaks the layout being redefined.

Fallout handled rather than left to fail quietly: all 22 hand-rolled test
packings go through the same helper production uses — as plain `u64` literals
they would have changed meaning without changing a line — and the config
product-ceiling test is rewritten, since `4 x profiled` is no longer reachable
exactly once `ssa_part_size` is capped at 255.

+++ crypto/packet/benches/packet_bench.rs
@@ -50,7 +50,7 @@ const PACKET_BENCHMARK: &[(usize, usize)] = &[
 ///
 /// This is what sets the per-share cost: `next_share` is a Horner evaluation over `threshold`
 /// coefficients.
-const PIX_THRESHOLD: u16 = DEFAULT_POLY_THRESHOLD;
+const PIX_THRESHOLD: u8 = DEFAULT_POLY_THRESHOLD;
 
 /// Whether the PIX share generator has a committed SSA.
 ///
